Fix SPARQL query: Assign expression to variable
SPARQL Grammar defines that in a SELECT-clause an expression should be
assigned to a variable:
    [9]   SelectClause    ::=   'SELECT' ( 'DISTINCT' | 'REDUCED' )? ( ( Var | ( '(' Expression 'AS' Var ')' ) )+ | '*' )
-- https://www.w3.org/TR/2013/REC-sparql11-query-20130321/#rSelectClause

Without assignment of `COUNT(*)` to a variable, Virtuoso is forgiving
and returns the result as `callret-0`. But mu-authorization terminates
with this error:

    (FunctionClauseError) no function clause matching in SparqlServer.Router.HandlerSupport.wrap_query_in_toplevel/1

This commit prevents the error by assigning to variable `triples`.

Clients should now expect the value in `triples`, not `callret-0`.